requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=2.0.0
tqdm>=4.66.0 